// uis/website/types/content.ts
export interface StatItem {
  value: string;
  label: string;
}

export interface BenefitItem {
  title: string;
  description: string;
}

export interface ServiceItem {
  title: string;
  description: string;
  image: string;
  alt: string;
}

export interface ContactItem {
  label: string;
  value: string;
}